+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.2.0] - 2018-09-12
 ### Added
 - `black`, `white`, `gray`, `red`, `green` and `blue` constructors to the `Color` struct.
 
@@ -30,5 +32,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - `Paint::from_span` poor performance.
 
-[Unreleased]: https://github.com/RazrFalcon/svgtypes/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/RazrFalcon/svgtypes/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/RazrFalcon/svgtypes/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/RazrFalcon/svgtypes/compare/v0.1.0...v0.1.1